[mod] [BREAKING] Simplify default error output
The goal of this commit is to improve the human readibility of
invariant violation errors.

The previous error output contained a lot of unnecessary visual noise,
sometimes obscuring the key information - especially for new users.

This commit will be BREAKING for the *very* small minority of users
that use Truss's (undocumented) error output programmatically. In this
case, please see the new output form to see what's changed.

All the same information is still present, but the structure and some
types have been cleaned up.