[mod] Make macros: `ex-info`, `ex-info!`, `unexpected-arg!`
Before this commit: `ex-info`, `ex-info!`, `unexpected-arg!` were functions
After  this commit: `ex-info`, `ex-info!`, `unexpected-arg!` are macros

This is technically a BREAKING change, though it's very unlikely
that anyone was actually using any of these in a way that'd be broken
by the change.

And by changing to macros, it became possible to include callsite
info ({:keys [ns coords]}) in resulting ex-data.

The callsite info can be handy, so I think it's worth the small
theoretical risk of a very small number of breakages.

Sincere apologies if anyone is broken by this!! If you are, it
hopefully shouldn't be difficult to update the affected code. Feel
free to ping me at taoensso.com/truss if I can assist.